Add view to list events per room
This commit adds the ability to see which events are happening in a
particular room. This is only reachable currently by navigating as
before until one reaches the event details view. On this view the room
has been made clickable and made to lead to the per-room event view. As
with the per-track event view, the spinner dropdown can be used to
select a different room.