Enable to list events by room and add speakers stub
This patch adds a "Rooms" button on the main view which allows the user
to view the events taking place in a particular room for a particilar
day. The day can be chosen from the drop down list at the top.

Also included is a disabled button to show speakers. This will be
enabled soon.